requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2025.4
Django==5.2.1
django-appconf==1.1.0
django-crispy-forms==2.4
django-filter==25.1
django-select2==8.4.1
sqlparse==0.5.3
